backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.43
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1
pytest==8.4.2
httpx==0.28.1
email-validator==2.3.0
feedparser==6.0.12
pydantic-ai==1.76.0
edge-tts==7.2.8
